Добавляет домашнее задание 'Модуляция'
